+++ ChangeLog.md
@@ -3,6 +3,7 @@
     - Added the permission system (Assignments and management)
     * Fixed poke and client description with empty message
     * Fixed channel tree icons
+    * Fixed group sorting
     
 * **26.09.18**:
     - Added Safari support